Add files via upload
This file takes an existing POSCAR file for a 4x4x1 sheet of graphene and replaces two carbons with one beryllium and one oxygen atom. The lattice point number for the replacements are specified. 

The base POSCAR file is read and the position information is reordered to reflect which atoms are replaced. A new POSCAR file is written that contains the appropriate positions. This is done iteratively. 

I would rather have my fingernails pulled out one by one than try and do this again. Python truly is the most rancid language available.